Modulated code so that ML operations and data preprocessing was seperated. This allowed for the diagnosis of the problem. Incorporated linear interpolation. Further debugging required.
