+++ lstm_model.py
@@ -3,9 +3,8 @@ from keras.models import Sequential
 from keras.layers import LSTM, Dense, Dropout, Bidirectional
 from keras.optimizers import Adam
 from sklearn.preprocessing import MinMaxScaler
-import sqlite3
 import pandas as pd
-from sklearn.preprocessing import StandardScaler
+from processor import prepare_data, fetch_stock_data_from_db, fetch_sentiment_data_from_db
 
 # Database configuration
 DB_NAME = 'data.sqlite'
@@ -14,87 +13,6 @@ DB_NAME = 'data.sqlite'
 stock_scaler = MinMaxScaler(feature_range=(0, 1))
 sentiment_scaler = MinMaxScaler(feature_range=(0, 1))
 look_back = 5
-
-
-
-def connect_to_db():
-    """
-        Establishes and returns a connection to the SQLite database.
-
-        Returns:
-            sqlite3.Connection: Database connection object.
-        """
-    return sqlite3.connect(DB_NAME)
-
-
-def fetch_stock_data_from_db():
-    conn = connect_to_db()
-    df = pd.read_sql('SELECT * FROM stock_data', conn)
-    conn.close()
-    return df
-
-
-def fetch_sentiment_data_from_db():
-    conn = connect_to_db()
-    df = pd.read_sql('SELECT date, overall_sentiment_score FROM sentiment_data', conn)
-    conn.close()
-    return df
-
-
-def prepare_data(stock_data, sentiment_data=None, look_back=5):
-    # Convert to DataFrame and set date as index
-    stock_df = pd.DataFrame(stock_data, columns=['Close'])
-    stock_df['date'] = pd.to_datetime(stock_df.index)
-    stock_df.set_index('date', inplace=True)
-
-    if sentiment_data is not None:
-        sentiment_df = pd.DataFrame(sentiment_data, columns=['overall_sentiment_score'])
-        sentiment_df['date'] = pd.to_datetime(sentiment_df.index)
-        sentiment_df.set_index('date', inplace=True)
-
-        # Check and handle initial NaNs in sentiment data
-        if sentiment_df['overall_sentiment_score'].isnull().iloc[0]:
-            # Filling initial NaNs with the first non-NaN value or a default value
-            first_valid_index = sentiment_df['overall_sentiment_score'].first_valid_index()
-            first_valid_value = sentiment_df['overall_sentiment_score'].loc[first_valid_index] if first_valid_index else 0
-            sentiment_df['overall_sentiment_score'].fillna(first_valid_value, inplace=True)
-
-        # Forward fill sentiment data
-        sentiment_df_ffill = sentiment_df.resample('D').ffill()
-
-        # Merge stock and sentiment data
-        merged_df = pd.merge(stock_df, sentiment_df_ffill, left_index=True, right_index=True, how='left')
-        merged_df.fillna(method='ffill', inplace=True)
-    else:
-        merged_df = stock_df
-
-    # Normalize the data
-    data_normalized = stock_scaler.fit_transform(merged_df[['Close']])
-
-    if 'overall_sentiment_score' in merged_df.columns:
-        # Print statistics and a small sample before normalization
-        print("Before normalization - Sentiment data statistics:\n", merged_df['overall_sentiment_score'].describe())
-        print("Before normalization - Sample of sentiment data:\n", merged_df['overall_sentiment_score'].head(5))
-
-        sentiment_normalized = sentiment_scaler.fit_transform(merged_df[['overall_sentiment_score']])
-        data_normalized = np.hstack((data_normalized, sentiment_normalized))
-
-        # Debugging: Check for NaNs after normalization
-        print("NaNs after normalization:", np.isnan(data_normalized).any())
-
-        # Print statistics and a small sample after normalization
-        print("After normalization - Sentiment data statistics:\n", pd.DataFrame(sentiment_normalized).describe())
-        print("After normalization - Sample of sentiment data:\n", sentiment_normalized[:5])
-
-
-    # Structure data for LSTM
-    X, y = [], []
-    for i in range(look_back, len(data_normalized)):
-        X.append(data_normalized[i - look_back:i])
-        y.append(data_normalized[i, 0])
-    X, y = np.array(X), np.array(y)
-
-    return X, y
 
 
 def train_lstm_model(X_train, y_train, epochs=50, batch_size=32):
@@ -139,15 +57,15 @@ def train_lstm_model(X_train, y_train, epochs=50, batch_size=32):
 
 def generate_plot_data():
     """
-        Generates data for plotting actual and predicted stock values.
+    Generates data for plotting actual and predicted stock values.
 
-        This function fetches stock and sentiment data from the database, preprocesses the data,
-        trains two LSTM models (one with sentiment and one without), makes predictions, and
-        prepares the data for plotting on the frontend.
+    This function fetches stock and sentiment data from the database, preprocesses the data,
+    trains two LSTM models (one with sentiment and one without), makes predictions, and
+    prepares the data for plotting on the frontend.
 
-        Returns:
-            list: List of dictionaries containing actual and predicted stock values.
-        """
+    Returns:
+        list: List of dictionaries containing actual and predicted stock values.
+    """
     # Fetch data from the database
     stock_df = fetch_stock_data_from_db()
     sentiment_df = fetch_sentiment_data_from_db()
@@ -165,6 +83,9 @@ def generate_plot_data():
     # Merge dataframes on date and fill NA values with the previous non-NA value (forward fill)
     merged_df = pd.merge(stock_df, sentiment_df_grouped, on='date', how='left').fillna(method='ffill')
 
+    # Fit the scaler on the original stock data
+    stock_scaler.fit(merged_df[['Close']])
+
     # Extract required columns for model processing
     stock_data = merged_df[['Close']].values
     sentiment_data = merged_df[['overall_sentiment_score']].values
@@ -181,13 +102,12 @@ def generate_plot_data():
     predicted_combined = model_combined.predict(X_combined)
     predicted_stock = model_stock.predict(X_stock)
 
-    # Rescale the predicted data to original scale
-    predicted_combined_rescaled = stock_scaler.inverse_transform(predicted_combined)
-    predicted_stock_rescaled = stock_scaler.inverse_transform(predicted_stock)
+    # Rescale the predicted data to original scale using the fitted scaler
+    predicted_combined_rescaled = stock_scaler.inverse_transform(predicted_combined.reshape(-1, 1))
+    predicted_stock_rescaled = stock_scaler.inverse_transform(predicted_stock.reshape(-1, 1))
     actual_data_rescaled = stock_scaler.inverse_transform(y_combined.reshape(-1, 1))
 
     # Use the dates from the merged dataframe as the x-values
-    # Since we're using look_back, we need to adjust the start date
     date_range = merged_df['date'][look_back:].values
 
     # Create data for plotting
